# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BaseRouteResponse": {
        "additionalProperties": false,
        "properties": {
          "distance_meters": {
            "minimum": 0.0,
            "title": "Distance Meters",
            "type": "number"
          },
          "duration_seconds": {
            "minimum": 0.0,
            "title": "Duration Seconds",
            "type": "number"
          },
          "geometry": {
            "$ref": "#/components/schemas/RouteGeometry"
          },
          "maneuvers": {
            "items": {
              "$ref": "#/components/schemas/ManeuverResponse"
            },
            "title": "Maneuvers",
            "type": "array"
          },
          "navigation": {
            "$ref": "#/components/schemas/NavigationTimingResponse"
          },
          "provider": {
            "const": "valhalla",
            "title": "Provider",
            "type": "string"
          },
          "speed_limit_source": {
            "anyOf": [
              {
                "const": "valhalla_graph",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Limit Source"
          },
          "speed_limits": {
            "items": {
              "$ref": "#/components/schemas/RouteSpeedLimitResponse"
            },
            "title": "Speed Limits",
            "type": "array"
          }
        },
        "required": [
          "distance_meters",
          "duration_seconds",
          "geometry",
          "maneuvers",
          "speed_limits",
          "speed_limit_source",
          "provider",
          "navigation"
        ],
        "title": "BaseRouteResponse",
        "type": "object"
      },
...
      "CoordinateRequest": {
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "CoordinateRequest",
        "type": "object"
      },
...
      "CorridorPolicyResponse": {
        "additionalProperties": false,
        "properties": {
          "cap_applied": {
            "enum": [
              "minimum",
              "maximum",
              "none"
            ],
            "title": "Cap Applied",
            "type": "string"
          },
          "effective_cng_range_km": {
            "exclusiveMinimum": 0.0,
            "title": "Effective Cng Range Km",
            "type": "number"
          },
          "radius_km": {
            "exclusiveMinimum": 0.0,
            "title": "Radius Km",
            "type": "number"
          },
          "range_fraction": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Range Fraction",
            "type": "number"
          },
          "uncapped_radius_km": {
            "exclusiveMinimum": 0.0,
            "title": "Uncapped Radius Km",
            "type": "number"
          }
        },
        "required": [
          "effective_cng_range_km",
          "range_fraction",
          "uncapped_radius_km",
          "radius_km",
          "cap_applied"
        ],
        "title": "CorridorPolicyResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "NetworkCostBasisResponse": {
        "additionalProperties": false,
        "properties": {
          "distance_model": {
            "const": "road_network",
            "title": "Distance Model",
            "type": "string"
          },
          "duration_model": {
            "enum": [
              "valhalla_graph_speeds",
              "valhalla_time_dependent_traffic"
            ],
            "title": "Duration Model",
            "type": "string"
          },
          "provider": {
            "const": "valhalla",
            "title": "Provider",
            "type": "string"
          },
          "traffic_aware": {
            "title": "Traffic Aware",
            "type": "boolean"
          },
          "traffic_state": {
            "enum": [
              "not_configured",
              "configured",
              "mock",
              "fresh",
              "stale",
              "unavailable"
            ],
            "title": "Traffic State",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "traffic_state",
          "traffic_aware",
          "duration_model",
          "distance_model"
        ],
        "title": "NetworkCostBasisResponse",
        "type": "object"
      },
      "NetworkEvaluationMetricsResponse": {
        "additionalProperties": false,
        "properties": {
          "base_route_calls": {
            "const": 1,
            "title": "Base Route Calls",
            "type": "integer"
          },
          "eligible_candidate_count": {
            "minimum": 0.0,
            "title": "Eligible Candidate Count",
            "type": "integer"
          },
          "excluded_by_detour_count": {
            "minimum": 0.0,
            "title": "Excluded By Detour Count",
            "type": "integer"
          },
          "matrix_batch_size": {
            "exclusiveMinimum": 0.0,
            "title": "Matrix Batch Size",
            "type": "integer"
          },
          "matrix_calls": {
            "minimum": 0.0,
            "title": "Matrix Calls",
            "type": "integer"
          },
          "matrix_candidate_count": {
            "minimum": 0.0,
            "title": "Matrix Candidate Count",
            "type": "integer"
          },
          "matrix_fallback_splits": {
            "description": "Binary batch splits used to isolate uncorrelatable locations.",
            "minimum": 0.0,
            "title": "Matrix Fallback Splits",
            "type": "integer"
          },
          "matrix_location_failures": {
            "description": "Isolated candidate locations with no suitable routing edge.",
            "minimum": 0.0,
            "title": "Matrix Location Failures",
            "type": "integer"
          },
          "per_candidate_route_calls": {
            "const": 0,
            "title": "Per Candidate Route Calls",
            "type": "integer"
          },
          "reachable_candidate_count": {
            "minimum": 0.0,
            "title": "Reachable Candidate Count",
            "type": "integer"
          },
          "spatial_candidate_count": {
            "minimum": 0.0,
            "title": "Spatial Candidate Count",
            "type": "integer"
          },
          "unreachable_candidate_count": {
            "minimum": 0.0,
            "title": "Unreachable Candidate Count",
            "type": "integer"
          }
        },
        "required": [
          "spatial_candidate_count",
          "matrix_candidate_count",
          "reachable_candidate_count",
          "unreachable_candidate_count",
          "eligible_candidate_count",
          "excluded_by_detour_count",
          "matrix_batch_size",
          "matrix_calls",
          "matrix_fallback_splits",
          "matrix_location_failures",
          "base_route_calls",
          "per_candidate_route_calls"
        ],
        "title": "NetworkEvaluationMetricsResponse",
        "type": "object"
      },
...
      "RankedCandidateResponse": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "destination_eta": {
            "format": "date-time",
            "title": "Destination Eta",
            "type": "string"
          },
          "detour_duration_seconds": {
            "minimum": 0.0,
            "title": "Detour Duration Seconds",
            "type": "number"
          },
          "detour_minutes": {
            "minimum": 0.0,
            "title": "Detour Minutes",
            "type": "number"
          },
          "distance_from_previous_waypoint_meters": {
            "description": "Road-network distance from the request origin/previous waypoint.",
            "minimum": 0.0,
            "title": "Distance From Previous Waypoint Meters",
            "type": "number"
          },
          "duration_from_previous_waypoint_seconds": {
            "minimum": 0.0,
            "title": "Duration From Previous Waypoint Seconds",
            "type": "number"
          },
          "extra_distance_meters": {
            "minimum": 0.0,
            "title": "Extra Distance Meters",
            "type": "number"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "mimit_station_id": {
            "minLength": 1,
            "title": "Mimit Station Id",
            "type": "string"
          },
          "municipality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipality"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "opening": {
            "$ref": "#/components/schemas/OpeningHoursEvaluationResponse"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "osm_match_confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Osm Match Confidence"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentCngPriceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "ranking": {
            "$ref": "#/components/schemas/RankingBreakdownResponse"
          },
          "route_fraction": {
            "description": "Approximate projection position from route start (0) to end (1).",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Route Fraction",
            "type": "number"
          },
          "route_via_station_distance_meters": {
            "minimum": 0.0,
            "title": "Route Via Station Distance Meters",
            "type": "number"
          },
          "route_via_station_duration_seconds": {
            "minimum": 0.0,
            "title": "Route Via Station Duration Seconds",
            "type": "number"
          },
          "station_eta": {
            "format": "date-time",
            "title": "Station Eta",
            "type": "string"
          },
          "station_id": {
            "exclusiveMinimum": 0.0,
            "title": "Station Id",
            "type": "integer"
          },
          "station_to_destination_distance_meters": {
            "minimum": 0.0,
            "title": "Station To Destination Distance Meters",
            "type": "number"
          },
          "station_to_destination_duration_seconds": {
            "minimum": 0.0,
            "title": "Station To Destination Duration Seconds",
            "type": "number"
          },
          "straight_line_distance_to_route_meters": {
            "description": "Cheap geographic prefilter distance. This is not road distance or detour distance.",
            "minimum": 0.0,
            "title": "Straight Line Distance To Route Meters",
            "type": "number"
          }
        },
        "required": [
          "station_id",
          "mimit_station_id",
          "name",
          "municipality",
          "province",
          "latitude",
          "longitude",
          "straight_line_distance_to_route_meters",
          "route_fraction",
          "distance_from_previous_waypoint_meters",
          "duration_from_previous_waypoint_seconds",
          "station_to_destination_distance_meters",
          "station_to_destination_duration_seconds",
          "route_via_station_distance_meters",
          "route_via_station_duration_seconds",
          "extra_distance_meters",
          "detour_duration_seconds",
          "detour_minutes",
          "station_eta",
          "destination_eta",
          "opening",
          "phone",
          "brand",
          "operator",
          "price",
          "ranking"
        ],
        "title": "RankedCandidateResponse",
        "type": "object"
      },
      "RankedCandidatesApiRequest": {
        "additionalProperties": false,
        "properties": {
          "costing": {
            "const": "auto",
            "default": "auto",
            "title": "Costing",
            "type": "string"
          },
          "departure_at": {
            "description": "Timezone-aware departure instant used to calculate station and destination ETAs.",
            "format": "date-time",
            "title": "Departure At",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/CoordinateRequest"
          },
          "effective_cng_range_km": {
            "description": "Usable CNG range after the caller's reserve/safety allowance. It controls spatial corridor width only.",
            "exclusiveMinimum": 0.0,
            "maximum": 2000.0,
            "title": "Effective Cng Range Km",
            "type": "number"
          },
          "include_closed": {
            "default": false,
            "description": "Include stations evaluated as closed at ETA with a zero opening score. Unknown opening state remains included regardless of this option.",
            "title": "Include Closed",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z]{2}(-[A-Za-z]{2})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "BCP 47 language tag used for maneuver instructions.",
            "title": "Language"
          },
          "maximum_detour_minutes": {
            "description": "Inclusive maximum added road-network travel time.",
            "maximum": 240.0,
            "minimum": 0.0,
            "title": "Maximum Detour Minutes",
            "type": "number"
          },
          "origin": {
            "$ref": "#/components/schemas/CoordinateRequest"
          },
          "origin_heading_degrees": {
            "anyOf": [
              {
                "exclusiveMaximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional observed travel heading at the origin. Used for a moving off-route recalculation so the new route departs in the current direction.",
            "title": "Origin Heading Degrees"
          },
          "origin_heading_tolerance_degrees": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Maximum angular difference accepted at the routed origin.",
            "title": "Origin Heading Tolerance Degrees"
          }
        },
        "required": [
          "origin",
          "destination",
          "departure_at",
          "effective_cng_range_km",
          "maximum_detour_minutes"
        ],
        "title": "RankedCandidatesApiRequest",
        "type": "object"
      },
      "RankedCandidatesResponse": {
        "additionalProperties": false,
        "properties": {
          "base_route": {
            "$ref": "#/components/schemas/BaseRouteResponse"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RankedCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "corridor": {
            "$ref": "#/components/schemas/CorridorPolicyResponse"
          },
          "cost_basis": {
            "$ref": "#/components/schemas/NetworkCostBasisResponse"
          },
          "departure_at": {
            "format": "date-time",
            "title": "Departure At",
            "type": "string"
          },
          "maximum_detour_minutes": {
            "minimum": 0.0,
            "title": "Maximum Detour Minutes",
            "type": "number"
          },
          "network_evaluation": {
            "$ref": "#/components/schemas/NetworkEvaluationMetricsResponse"
          },
          "ranking_evaluation": {
            "$ref": "#/components/schemas/RankingMetricsResponse"
          },
          "ranking_policy": {
            "$ref": "#/components/schemas/RankingPolicyResponse"
          },
          "spatial_pruning": {
            "$ref": "#/components/schemas/SpatialPruningMetricsResponse"
          },
          "stage": {
            "const": "ranking",
            "default": "ranking",
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "departure_at",
          "maximum_detour_minutes",
          "base_route",
          "corridor",
          "spatial_pruning",
          "cost_basis",
          "network_evaluation",
          "ranking_policy",
          "ranking_evaluation",
          "candidates"
        ],
        "title": "RankedCandidatesResponse",
        "type": "object"
      },
...
      "RankingMetricsResponse": {
        "additionalProperties": false,
        "properties": {
          "detour_eligible_candidate_count": {
            "minimum": 0.0,
            "title": "Detour Eligible Candidate Count",
            "type": "integer"
          },
          "enrichment_queries": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Enrichment Queries",
            "type": "integer"
          },
          "excluded_closed_count": {
            "minimum": 0.0,
            "title": "Excluded Closed Count",
            "type": "integer"
          },
          "opening_closed_count": {
            "minimum": 0.0,
            "title": "Opening Closed Count",
            "type": "integer"
          },
          "opening_invalid_count": {
            "minimum": 0.0,
            "title": "Opening Invalid Count",
            "type": "integer"
          },
          "opening_missing_count": {
            "minimum": 0.0,
            "title": "Opening Missing Count",
            "type": "integer"
          },
          "opening_open_count": {
            "minimum": 0.0,
            "title": "Opening Open Count",
            "type": "integer"
          },
          "opening_unknown_count": {
            "minimum": 0.0,
            "title": "Opening Unknown Count",
            "type": "integer"
          },
          "opening_valid_count": {
            "minimum": 0.0,
            "title": "Opening Valid Count",
            "type": "integer"
          },
          "price_available_count": {
            "minimum": 0.0,
            "title": "Price Available Count",
            "type": "integer"
          },
          "price_missing_count": {
            "minimum": 0.0,
            "title": "Price Missing Count",
            "type": "integer"
          },
          "ranked_candidate_count": {
            "minimum": 0.0,
            "title": "Ranked Candidate Count",
            "type": "integer"
          }
        },
        "required": [
          "detour_eligible_candidate_count",
          "opening_open_count",
          "opening_closed_count",
          "opening_unknown_count",
          "opening_valid_count",
          "opening_missing_count",
          "opening_invalid_count",
          "excluded_closed_count",
          "price_available_count",
          "price_missing_count",
          "ranked_candidate_count",
          "enrichment_queries"
        ],
        "title": "RankingMetricsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OptionalHTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CNG-aware navigation API foundations for Italy.",
    "title": "Compass CNG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cng/ranked-candidates": {
      "post": {
        "operationId": "ranked_candidates_api_v1_cng_ranked_candidates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankedCandidatesApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedCandidatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or no route."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid routing provider response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or routing unavailable."
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Ranked Candidates",
        "tags": [
          "ranking"
        ]
      }
    },
...
  }
}